port: Fix typo in pthread type definition and remove Windows-specific includes for Linux compatibility

+++ src/thread_args.h
@@ -1,10 +1,8 @@
 #ifndef THREAD_ARGS_H
 #define THREAD_ARGS_H
 
-#include <windows.h>
-
 typedef struct {
-    DWORD sleep_time;
+    unsigned long sleep_time;
     int dialog_n;
     int dialog_shown;
     volatile int running;
